Moving all pipeline stuff into a contrib directory.
